Port basic edit page to `judgments` app from `judgments_admin`
Remove old `judgments_admin` app
