+++ config/settings/base.py
@@ -79,8 +79,7 @@ THIRD_PARTY_APPS = [
 
 LOCAL_APPS = [
     # Your stuff: custom apps go here
-    "judgments.apps.JudgmentsConfig",
-    "judgments_admin.apps.JudgmentsAdminConfig",
+    "judgments.apps.JudgmentsConfig"
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
