+++ ds_caselaw_editor_ui/templates/includes/basic_search_form.html
@@ -13,9 +13,5 @@
       </div>
 
     </form>
-
-    <div class="search-component__more-options">
-      <a href="{% url 'structured_search' %}">More search options</a>
-    </div>
   </div>
 </div>
