+++ ds_caselaw_editor_ui/templates/includes/judgment_text_toolbar.html
@@ -6,8 +6,12 @@
                 <a href="{{ request.args.return_link }}">{% translate "judgment.back" %}</a>
             {% endif %}
         </div>
-        <div class="judgment-toolbar__download">
-            <a class="judgment-download__option--xml" href="{% url 'detail_xml' context.judgment_uri %}">{% translate "judgment.downloadasxml" %}</a>
+        <div class="judgment-toolbar__edit">
+          {% if context.judgment  %}
+            <a class="" href="{% url 'edit' %}?judgment_uri={{context.judgment_uri}}">Edit this Judgment</a>
+          {% elif request.path == "/edit" %}
+            <a class="" href="{% url 'detail' %}?judgment_uri={{context.judgment_uri}}">View this Judgment</a>
+          {% endif %}
         </div>
     </div>
 </div>
