+++ ds_caselaw_editor_ui/templates/judgment/results.html
@@ -16,12 +16,9 @@
   <div class="results">
     <h1>{{ pagetitle }}</h1>
 
-    {% include 'includes/results_search_component.html' %}
+    {% include 'includes/basic_search_form.html' %}
     <div class="results__container">
       <div class="results__list-layout-container">
-        <div class="results__result-controls-container">
-          {% include 'includes/result_controls.html' %}
-        </div>
         <div class="results__result-list-container">
           {% include 'includes/results_list.html' %}
           {% include 'includes/pagination.html' %}
